Fix bug in EmitCopy affecting certain match lengths.
EmitCopy caps match lengths at 65535 bytes as needed by EmitCopyGreaterThan63,
but only when over 65539. This causes compression failures for match lengths
of 65536, 65537, and 65538.

Change it to ensure we always cap lengths, and that the remainder, if any, is
at least 4 bytes for EmitCopyLessThan63.

The below code triggers the problem by compressing 65537 zeros, which will
become a single literal and a copy of length 65536:

    size_t n = 65537, m = pithy_MaxCompressedLength(n);
    char *p = calloc(n, 1), *q = malloc(m);
    pithy_Compress(p, n, q, m, 0);